day 11: day 2 solution, way faster
In that it will complete before it eats hundreds of gigs of memory on my
machine. Most of the numbers we're calculating here are repeats, because
we end up in a loop for nearly everything. So all we need to do is keep
a count of how many (say) 8s we have in the list now, and compute the
next step for 8 once add that many copies of the results instead.